refactor(CollectionSettingsPage): remove update_collection command and implement new tab components for authentication, headers, variables, and scripts
