+++ view/desktop/bin/src/lib.rs
@@ -153,7 +153,6 @@ pub async fn run<R: TauriRuntime>() {
             commands::describe_workspace_state,
             commands::stream_collections,
             commands::create_collection,
-            commands::update_collection,
             commands::delete_collection,
             //
             // Collection
