+++ view/desktop/src/hooks/collection/useUpdateCollection.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { invokeTauriIpc } from "@/lib/backend/tauri";
-import { UpdateCollectionInput, UpdateCollectionOutput } from "@repo/moss-workspace";
+// import { invokeTauriIpc } from "@/lib/backend/tauri";
+// import { UpdateCollectionInput, UpdateCollectionOutput } from "@repo/moss-workspace";
 import { useCollectionsStore } from "@/store/collections";
 
 interface UpdateCollectionParams {
@@ -16,22 +16,36 @@ export const useUpdateCollection = () => {
 
   return useMutation({
     mutationFn: async ({ collectionId, name, order, pinned }: UpdateCollectionParams) => {
-      const input: UpdateCollectionInput = {
-        id: collectionId,
-        newName: name || null,
-        order,
-        pinned,
+      // Mock implementation - simulate network delay
+      await new Promise((resolve) => setTimeout(resolve, 200));
+
+      // TODO: Replace with actual backend call when update_collection is implemented
+      // const input: UpdateCollectionInput = {
+      //   id: collectionId,
+      //   newName: name || undefined,
+      //   order,
+      //   pinned,
+      // };
+      //
+      // const result = await invokeTauriIpc<UpdateCollectionOutput>("update_collection", { input });
+      //
+      // if (result.status === "error") {
+      //   throw new Error(String(result.error));
+      // }
+      //
+      // return result.data;
+
+      // Mock successful response
+      return {
+        collection: {
+          id: collectionId,
+          name: name || `Collection ${collectionId}`,
+          order: order || 0,
+          pinned: pinned || false,
+        },
       };
-
-      const result = await invokeTauriIpc<UpdateCollectionOutput>("update_collection", { input });
-
-      if (result.status === "error") {
-        throw new Error(String(result.error));
-      }
-
-      return result.data;
     },
-    onSuccess: (data, variables) => {
+    onSuccess: (_data, variables) => {
       // Update the store if we have the collection data
       const collections = useCollectionsStore.getState().collections;
       const targetCollection = collections.find(
