+++ view/desktop/src/pages/CollectionSettingsPage/tabs/AuthTabContent.tsx
