+++ view/desktop/src/pages/CollectionSettingsPage/tabs/HeadersTabContent.tsx
