+++ view/desktop/src/pages/CollectionSettingsPage/tabs/VariablesTabContent.tsx
